+++ test/inputs/collection_select_input_test.rb
@@ -31,7 +31,7 @@ class CollectionSelectInputTest < ActionView::TestCase
   test 'input should do automatic collection translation for select types using defaults key' do
     store_translations(:en, simple_form: { options: { defaults: {
       gender: { male: 'Male', female: 'Female'}
-    } } } ) do
+    } } }) do
       with_input_for @user, :gender, :select, collection: [:male, :female]
       assert_select 'select.select#user_gender'
       assert_select 'select option', 'Male'
@@ -42,7 +42,7 @@ class CollectionSelectInputTest < ActionView::TestCase
   test 'input should do automatic collection translation for select types using specific object key' do
     store_translations(:en, simple_form: { options: { user: {
       gender: { male: 'Male', female: 'Female'}
-    } } } ) do
+    } } }) do
       with_input_for @user, :gender, :select, collection: [:male, :female]
       assert_select 'select.select#user_gender'
       assert_select 'select option', 'Male'
@@ -102,9 +102,63 @@ class CollectionSelectInputTest < ActionView::TestCase
     assert_select 'select option[value=]', ''
   end
 
+  test 'input should translate include blank when set to :translate' do
+    store_translations(:en, simple_form: { include_blanks: { user: {
+      age: 'Rather not say'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, include_blank: :translate
+      assert_select 'select option[value=]', 'Rather not say'
+    end
+  end
+
+  test 'input should translate include blank with a default' do
+    store_translations(:en, simple_form: { include_blanks: { defaults: {
+      age: 'Rather not say',
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, include_blank: :translate
+      assert_select 'select option[value=]', 'Rather not say'
+    end
+  end
+
+  test 'input should not translate include blank when set to a string' do
+    store_translations(:en, simple_form: { include_blanks: { user: {
+      age: 'Rather not say'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, include_blank: 'Young at heart'
+      assert_select 'select option[value=]', 'Young at heart'
+    end
+  end
+
+  test 'input should not translate include blank when automatically set' do
+    store_translations(:en, simple_form: { include_blanks: { user: {
+      age: 'Rather not say'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30
+      assert_select 'select option[value=]', ''
+    end
+  end
+
+  test 'input should not translate include blank when set to true' do
+    store_translations(:en, simple_form: { include_blanks: { user: {
+      age: 'Rather not say'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, include_blank: true
+      assert_select 'select option[value=]', ''
+    end
+  end
+
+  test 'input should not translate include blank when set to false' do
+    store_translations(:en, simple_form: { include_blanks: { user: {
+      age: 'Rather not say'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, include_blank: false
+      assert_no_select 'select option[value=]'
+    end
+  end
+
   test 'input should not set include blank if otherwise is told' do
     with_input_for @user, :age, :select, collection: 18..30, include_blank: false
-    assert_no_select 'select option[value=]', ''
+    assert_no_select 'select option[value=]'
   end
 
   test 'input should not set include blank if prompt is given' do
@@ -115,6 +169,51 @@ class CollectionSelectInputTest < ActionView::TestCase
   test 'input should not set include blank if multiple is given' do
     with_input_for @user, :age, :select, collection: 18..30, input_html: { multiple: true }
     assert_no_select 'select option[value=]', ''
+  end
+
+  test 'input should translate prompt when set to :translate' do
+    store_translations(:en, simple_form: { prompts: { user: {
+      age: 'Select age:'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, prompt: :translate
+      assert_select 'select option[value=]', 'Select age:'
+    end
+  end
+
+  test 'input should translate prompt with a default' do
+    store_translations(:en, simple_form: { prompts: { defaults: {
+      age: 'Select age:',
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, prompt: :translate
+      assert_select 'select option[value=]', 'Select age:'
+    end
+  end
+
+  test 'input should not translate prompt when set to a string' do
+    store_translations(:en, simple_form: { prompts: { user: {
+      age: 'Select age:'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, prompt: 'Do it:'
+      assert_select 'select option[value=]', 'Do it:'
+    end
+  end
+
+  test 'input should not translate prompt when set to false' do
+    store_translations(:en, simple_form: { prompts: { user: {
+      age: 'Select age:'
+    } } }) do
+      with_input_for @user, :age, :select, collection: 18..30, prompt: false
+      assert_no_select 'select option[value=]'
+    end
+  end
+
+  test 'input should use Rails prompt translation as a fallback' do
+    store_translations(:en, helpers: { select: {
+      prompt: 'Select value:'
+    } }) do
+      with_input_for @user, :age, :select, collection: 18..30, prompt: :translate
+      assert_select 'select option[value=]', "Select value:"
+    end
   end
 
   test 'input should detect label and value on collections' do